Fixes wrong use of WindowsVersion
The original Utils.IsVista() and Utils.IsXp()  would return True ot False if the OS was greater than or eqiual to. We had changed everything to use the new WindowsVersion.IsVista() and IsXP() which would only return True/False if it is that opperating system.

This fixes all the wrongly used IsVista() and IsXP() to use WindowsVersion >= 'Vista' and WindowsVersion >= 'XP'